API Refactor: carto.dataview (update references)
