+++ src/lib/widgets/formula/Formula.ts
@@ -1,4 +1,4 @@
-import { Formula as FormulaDataView } from '@/dataviews/formula/Formula';
+import { FormulaDataView } from '@/dataview/formula/FormulaDataView';
 import { Widget } from '../widget';
 
 export class Formula extends Widget {
